changed the name of the file and made long long int to int

+++ 210595_210316_A2.c
@@ -1,16 +1,17 @@
 #include <stdio.h>
 #include <stdlib.h>
 
+// defining linked list
 struct LinkedList
 {
-    long long int index_x, index_y, value;
+    int index_x, index_y, value;
     struct LinkedList *right;
     struct LinkedList *bottom;
 };
 typedef struct LinkedList *node;
 
 //  making connections in right and bottom direction
-node CreateNode(node row_pointer[], node col_pointer[], long long int index_x, long long int index_y, long long int value, node prev_node_x, node prev_node_y)
+node CreateNode(node row_pointer[], node col_pointer[], int index_x, int index_y, int value, node prev_node_x, node prev_node_y)
 {
 
     node temp = (node)malloc(sizeof(struct LinkedList));
@@ -19,7 +20,8 @@ node CreateNode(node row_pointer[], node col_pointer[], long long int index_x, l
     temp->index_y = index_y;
     temp->right = NULL;
     temp->bottom = NULL;
-    //creating linked list based matrix 
+
+    // creating linked list based matrix
 
     if (col_pointer[index_y] != NULL)
     {
@@ -46,18 +48,21 @@ node CreateNode(node row_pointer[], node col_pointer[], long long int index_x, l
 int main()
 {
     // Matrix Formation
-    long long int n, matrix_no, row, col, k, value, flag = 0, check = 1;
+    int n, matrix_no, row, col, k, value, flag = 0, check = 1;
+
     // flag is there to check if the matrix formed is NULL Matrix or not
-    scanf("%lld", &n);
+    scanf("%d", &n);
     node m1_row_pointer[n + 1], m1_col_pointer[n + 1];
     node m2_row_pointer[n + 1], m2_col_pointer[n + 1];
+
     // row_pointer and col_pointer are pointing at the row and column starting of the linked list
     node m1_prev_node_x[n + 1], m1_prev_node_y[n + 1];
     node m2_prev_node_x[n + 1], m2_prev_node_y[n + 1];
+
     // pre node pointers are used to create linked list based data structure pointing in bottom and right direction
-    for (long long int i = 0; i < n + 1; i++)
-    {  
-        //initializing all pointers as NULL
+    for (int i = 0; i < n + 1; i++)
+    {
+        // initializing all pointers as NULL
         m1_row_pointer[i] = NULL;
         m1_col_pointer[i] = NULL;
 
@@ -73,10 +78,10 @@ int main()
 
     while (check)
     {
-        for (long long int i = 0; i < 4; i++)
+        for (int i = 0; i < 4; i++)
         {
-            // scanning all the 4 datas in a given line
-            scanf("%lld", &k);
+            // scanning all the 4 datas in a given line and adjusting check if 0 is scanned
+            scanf("%d", &k);
             if (k == 0)
             {
                 check = 0;
@@ -118,16 +123,18 @@ int main()
     }
 
     // Matrix Multiplication
-    for (long long int i = 1; i <= n; i++)
+    for (int i = 1; i <= n; i++)
     {
         if (m1_row_pointer[i] == NULL)
             continue; // To skip null rows
-        for (long long int j = 1; j <= n; j++)
+
+        for (int j = 1; j <= n; j++)
         {
             if (m2_col_pointer[j] == NULL)
                 continue; // To skip null columns
+
             node temp_row_pointer = m1_row_pointer[i], temp_col_pointer = m2_col_pointer[j];
-            long long int sum = 0;
+            int sum = 0;
             while (temp_row_pointer != NULL && temp_col_pointer != NULL)
             {
                 if ((temp_row_pointer->index_y) == (temp_col_pointer->index_x))
@@ -153,7 +160,7 @@ int main()
             }
             if (sum != 0)
             {
-                printf("%lld %lld %lld\n", i, j, sum);
+                printf("%d %d %d\n", i, j, sum);
                 flag = 1;
             }
         }
